Add files via upload
A film recommender function based on the similarities of the ratings between users. The similarities are calculated with the KNNBasic of the surprise library (installing scikit-surprise is necessary).  
The KNNBasic is an algorithm that predicts a rating that a user would be giving a film considering distances (similar to K-Nearest Neighbors).